requirements.txt
alembic==1.7.4
attrs==21.2.0
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
SQLAlchemy==1.4.26
toml==0.10.2
Werkzeug==2.0.2
